Add API Endpoints for ItemQuality and InventoryTypes
