+++ WarcraftGearPlanner.Server/Services/InventoryTypes/IInventoryTypeService.cs
@@ -1,0 +1,8 @@
+﻿using WarcraftGearPlanner.Server.Data.Entities;
+using WarcraftGearPlanner.Shared.Models.Items;
+
+namespace WarcraftGearPlanner.Server.Services.InventoryTypes;
+
+public interface IInventoryTypeService : IService<InventoryType, InventoryTypeEntity>
+{
+}
