+++ WarcraftGearPlanner.Server/Services/InventoryTypes/InventoryTypeValidator.cs
@@ -1,0 +1,12 @@
+﻿using FluentValidation;
+using WarcraftGearPlanner.Shared.Models.Items;
+
+namespace WarcraftGearPlanner.Server.Services.InventoryTypes;
+
+public class InventoryTypeValidator : AbstractValidator<InventoryType>
+{
+	public InventoryTypeValidator()
+	{
+
+	}
+}
