+++ WarcraftGearPlanner.Server/Services/ItemClasses/IItemClassService.cs
@@ -1,7 +1,7 @@
 ﻿using WarcraftGearPlanner.Server.Data.Entities;
 using WarcraftGearPlanner.Shared.Models.Items;
 
-namespace WarcraftGearPlanner.Server.Services.Items;
+namespace WarcraftGearPlanner.Server.Services.ItemClasses;
 
 public interface IItemClassService : IService<ItemClass, ItemClassEntity>
 {
