+++ WarcraftGearPlanner.Server/Services/ItemQualities/IItemQualityService.cs
@@ -1,0 +1,8 @@
+﻿using WarcraftGearPlanner.Server.Data.Entities;
+using WarcraftGearPlanner.Shared.Models.Items;
+
+namespace WarcraftGearPlanner.Server.Services.ItemQualities;
+
+public interface IItemQualityService : IService<ItemQuality, ItemQualityEntity>
+{
+}
