+++ WarcraftGearPlanner.Server/Services/ItemQualities/ItemQualityValidator.cs
@@ -1,0 +1,12 @@
+﻿using FluentValidation;
+using WarcraftGearPlanner.Shared.Models.Items;
+
+namespace WarcraftGearPlanner.Server.Services.ItemQualities;
+
+public class ItemQualityValidator : AbstractValidator<ItemQuality>
+{
+	public ItemQualityValidator()
+	{
+
+	}
+}
